Fix: LSTM에 필요한 데이터 및 로직 수정

+++ app/src/main/java/com/example/myapplication/domain/HealthRequest.kt
@@ -8,7 +8,7 @@ data class HealthRequest (
     val walkingSteps: Int,
     val totalCaloriesBurned: Double,
     val spo2: Int,
-    val heartRateAvg: Double,
+    val heartRateAvg: Long,
 
     // 수면 단계 필드
     val sleepDurationMin: Long,
